+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Improvement:
 - [#1549](https://github.com/elysiajs/elysia/pull/1549) Set-Cookie headers not sent when errors are thrown
 - [#1579](https://github.com/elysiajs/elysia/pull/1579) HEAD to handle Promise value
 
+Security:
+- cookie injection, prototype pollution, and RCE
+
 Bug fix:
 - [#1550](https://github.com/elysiajs/elysia/pull/1550) excess property checking